+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Support numerical keys with Couchbase
 - Fix how Couchbase (failed to) handle negative TTLs
 - Make Memcached delete-by-expiration more reliable
+- Make HHVM work with numerical cache keys in Memcached::setMulti
 
 
 ## [1.4.2] - 2017-02-07